+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 - [Changelog](#changelog)
-  - [v0.0.42(2023-05-XX)](#v00422023-05-xx)
+  - [Next v0.0.42(2023-05-XX)](#next-v00422023-05-xx)
   - [v0.0.41(2023-05-22)](#v00412023-05-22)
   - [v0.0.40(2023-05-09)](#v00402023-05-09)
   - [v0.0.39(2023-05-01)](#v00392023-05-01)
@@ -22,16 +22,27 @@
   - [0.0.24(2023-02-24)](#00242023-02-24)
   - [\<= 0.0.23(2023-02-17)](#-00232023-02-17)
 
-## v0.0.42(2023-05-XX)
+## Next v0.0.42(2023-05-XX)
 
-- feat: `func` yellow light bulb `add ahkDoc`
+- api-change
+  1. del `"AhkNekoHelp.customize.signatureHelp"`
+  2. add
+
+  ```jsonc
+  {
+      "AhkNekoHelp.signatureHelp.InsertType": true, // false
+      "AhkNekoHelp.signatureHelp.showParamInfo": true, // false
+      "AhkNekoHelp.signatureHelp.showOtherDoc": true, // false
+      "AhkNekoHelp.signatureHelp.ReturnStyle": "auto" // or "allowShow" or "never"
+  }
+  ```
+
+- feat: `func` generate `ahkDoc`
   >
   > 1. api-change `@Return` -> `@returns`
   > 2. `@param` can generate type from default
      > ![img](https://raw.githubusercontent.com/CoffeeChaton/vscode-autohotkey-NekoHelp/main/image/func_yellow_light_bulb_add_ahkDoc2.gif)
 
-- api-change
-  1. del `"AhkNekoHelp.customize.signatureHelp"` options `0`
 - feat: add `"AhkNekoHelp.customize.signatureHelpInsertType"`, [4k 36 sec](https://youtu.be/HxjzRf7stpA)
 - fix: syntax-highlight of `,ByRef` or `)or`
 - feat: add `Return` -> `func` def, if `func` not has `label:` then support Go-to-func-Definition on `return` Keywords